requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
Django==5.0.3
django-bootstrap4==24.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-filter==24.1
django-mathfilters==1.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
environs==11.0.0
idna==3.7
marshmallow==3.21.2
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
